requirements.txt

flask==2.3.3
nltk==3.8.1
numpy==1.25.2
networkx==3.1
transformers==4.33.2
torch==2.0.1
spacy==3.6.1
python-docx==0.8.11
PyPDF2==3.0.1
Werkzeug==2.3.7